+++ src/Header/Request/Range.php
@@ -50,6 +50,14 @@ final class Range implements Header
     }
 
     /**
+     * @return true The header is always valid.
+     */
+    public function isValid(): bool
+    {
+        return true;
+    }
+
+    /**
      * @inheritDoc
      */
     public function getName(): string