+++ src/Header/Response/ContentDisposition.php
@@ -80,6 +80,14 @@ final class ContentDisposition implements Header
     }
 
     /**
+     * @return true The header is always valid.
+     */
+    public function isValid(): bool
+    {
+        return true;
+    }
+
+    /**
      * @inheritDoc
      */
     public function getName(): string