+++ src/Header/Response/ETag.php
@@ -49,6 +49,14 @@ final class ETag implements Header
     }
 
     /**
+     * @return true The header is always valid.
+     */
+    public function isValid(): bool
+    {
+        return true;
+    }
+
+    /**
      * @inheritDoc
      */
     public function getName(): string